first version to include furnace control

+++ controllers/databasecontroller.js
@@ -2,10 +2,12 @@
 
 const mysql = require('mysql2');
 var comController = require ('./communicationsController');
+const recircCNTRL = require ('./recircController');
+const furnaceController = require ('./furnaceController');
 
 // variables for the temperature sensor
 var tempSum1 = 0;
-var tempSum2 = 0;
+var tempFamSum = 0;
 var tempSum3 = 0;
 var tempSum4 = 0;
 var tempSum5 = 0;
@@ -13,7 +15,7 @@ var tempSum6 = 0;
 var tempSum7 = 0;
 var tempSum8 = 0;
 var avgTemp1 = 0;
-var avgTemp2 = 0;
+var avgFamTemp = 0;
 var avgTemp3 = 0;
 var avgTemp4 = 0;
 var avgTemp5 = 0;
@@ -21,7 +23,8 @@ var avgTemp6 = 0;
 var avgTemp7 = 0;
 var avgTemp8 = 0;
 var flags = [];
-var numOfReadingsToAvg = 10;
+//var numOfReadingsToAvg = 20;
+var numOfReadingsToAvg = 5;
 var tempcount = 0;
 //  30 min * 60 * numOfReadingsToAvg = 1800 - save every 30 min
 //  2  min * 60 * numOfReadingsToAvg / 2 = 600
@@ -32,14 +35,18 @@ var tempcount = 0;
 //  2 sec * 10 * 90 / 60 = 30 Min - current time between saves
 //var currentSaveDelayCount = 90;
 // the period between all temperature saves
-var saveDelayIntervalMinutes = 1;
+var saveDelayIntervalMinutes = .2;
 var saveDelayIntervalSeconds = saveDelayIntervalMinutes * 60;
 var currentSaveDelayCount = saveDelayIntervalSeconds;
-var currentDelayCountMin = 0; //var saveDelay = currentSaveDelayCount;
+var currentDelayCountMin = 0; // only used to display the count
+//var saveDelay = currentSaveDelayCount;
 //var delayCount = 0;
 var numDataPointsRead = 40;
 var temporaryTimes = [];
 var test = "Test";
+
+var furnChangeState = 'NULL';
+var localFurnAction = "noChange";
 
 var connection;
 
@@ -53,8 +60,9 @@ var myVar = setInterval(myTimer, 1000);
 
 //var saveDelayInSec = currentSaveDelayCount * 60;
 function myTimer(){
-  currentSaveDelayCount = currentSaveDelayCount - 1;
+  currentSaveDelayCount--;
   currentDelayCountMin = secondsToHms(currentSaveDelayCount);
+  console.log("current delay count - " + currentSaveDelayCount);
   console.log("current delay count - " + currentDelayCountMin);
 };
 
@@ -66,10 +74,14 @@ function secondsToHms(d) {
     return m + ":" + s; 
 };
 
+exports.upDateFurnState = function (newState){
+  furnChangeState = newState;
+}
+
 exports.upDateTempSaveIterval = function(newDelay) {
   saveDelayIntervalMinutes = newDelay;
   saveDelayIntervalSeconds = saveDelayIntervalMinutes * 60;
-  return ("Saved New Temp Save INterval");
+  return ("Saved New Temp Save Interval");
 };
 
 exports.updateNumDataPointsToChart = function(newDataPoints) {
@@ -120,7 +132,7 @@ connection.connect((err) => {
     var oursql = temp.concat(numDataPointsRead);
     //console.log(oursql);
     connection.query( oursql, (err, result) => {
-        console.log(result);
+        //console.log(result);
         if (err) {
           console.log("Got a DB error in getTempData (other)");
           console.log (err);
@@ -153,18 +165,36 @@ connection.connect((err) => {
     return dbSettings;
   };
 
-  exports.saveTempData = function (temp1, temp2, temp3, temp4, temp5, temp6, temp7) {
-    console.log("in save temperature data");
-    console.log(temp1, temp2, temp3, temp4, temp5, temp6, temp7);
-    //testSaveDelayInterval = testSaveDelayInterval - 2;
-    console.log(currentDelayCountMin + " - current save delay Count");
-    //console.log( "tempcount - " + tempcount + " Current Save Delay Count - " + currentSaveDelayCount);
+  exports.setFurnaceChange = function (furnChangeText){
+    localFurnAction = furnChangeText;
+  };
 
-    if (currentSaveDelayCount < 23){
+  exports.getCurentAvgTemps = function (){
+    var dataPac = [avgTemp1, avgFamTemp, avgTemp3, avgTemp4, avgTemp5, avgTemp6, avgTemp7];
+        // 4 -  tempF1 Wood Stove
+        // 5 -  tempF2 bread Board family room
+        // 6 -  tempF3 bedroom
+        // 7 -  tempF4 pipe
+        // 8 -  tempFurnaceF Furnace
+        // 9 -  tempF6 breadboard
+        // 10 - tempF7 outdoor sun
+    return (dataPac);
+  };
+
+
+  exports.saveTempData = function (temp1, temp2, temp3, temp4, temp5, temp6, temp7, furnAction) {
+    console.log("in save temperature data");
+    console.log(temp1, temp2, temp3, temp4, temp5, temp6, temp7, furnAction);
+    //testSaveDelayInterval = testSaveDelayInterval - 2;
+    currentSaveDelayCount--;
+    console.log(currentSaveDelayCount + " - current save delay Count");
+
+    //console.log( "tempcount - " + tempcount + " Current Save Delay Count - " + currentSaveDelayCount);
+    //if (currentSaveDelayCount < 23){
 
         // average temperature readings to numOfReadingsToAvg
         tempSum1 = tempSum1 + temp1;
-        tempSum2 = tempSum2 + temp2;
+        tempFamSum = tempFamSum + temp2;
         tempSum3 = tempSum3 + temp3;
         tempSum4 = tempSum4 + temp4;
         tempSum5 = tempSum5 + temp5;
@@ -174,27 +204,30 @@ connection.connect((err) => {
 
         if (tempcount == numOfReadingsToAvg) {
           avgTemp1 = parseFloat((tempSum1/tempcount).toFixed(1));
-          avgTemp2 = parseFloat((tempSum2/tempcount).toFixed(1));
+          avgFamTemp = parseFloat((tempFamSum/tempcount).toFixed(1));
           avgTemp3 = parseFloat((tempSum3/tempcount).toFixed(1));
           avgTemp4 = parseFloat((tempSum4/tempcount).toFixed(1));
           avgTemp5 = parseFloat((tempSum5/tempcount).toFixed(1));
           avgTemp6 = parseFloat((tempSum6/tempcount).toFixed(1));
           avgTemp7 = parseFloat((tempSum7/tempcount).toFixed(1));
           tempSum1 = 0;
-          tempSum2 = 0;
+          tempFamSum = 0;
           tempSum3 = 0;
           tempSum4 = 0;
           tempSum5 = 0;
           tempSum6 = 0;
           tempSum7 = 0;
           tempcount = 0;
-          console.log("averages - " + avgTemp1, avgTemp2, avgTemp3, avgTemp4, avgTemp5, avgTemp6, avgTemp7);
+          console.log("averages - " + avgTemp1, avgFamTemp, avgTemp3, avgTemp4, avgTemp5, avgTemp6, avgTemp7);
+
+          furnaceController.checkFurnace(avgTemp5, avgFamTemp, avgTemp3, avgTemp6, avgTemp7);
 
     //      console.log("Current Save Delay Count - " + currentSaveDelayCount);
-//          if (currentSaveDelayCount == 0){
+          if (currentSaveDelayCount <= 0){
 
     //        comController.returnFlags = function(flags){
     //          console.log("Flags in db controller - " +  flags)
+
               console.log("Saving Temp Data");
               connection.query("delete from temperatures ORDER BY id limit 1", (err) => {
                 if (err) {
@@ -208,12 +241,13 @@ connection.connect((err) => {
                 {
                   tempOutDoorsSun: avgTemp7,
                   tempOutDoorsShade: 40,
-                  tempFamilyRoom: avgTemp2,
+                  tempFamilyRoom: avgFamTemp,
                   tempBedRoom: avgTemp3,
                   tempDesk: avgTemp6,
                   tempPipe: avgTemp4,
                   tempWoodStove: avgTemp1,
-                  tempFurnace: avgTemp5
+                  tempFurnace: avgTemp5,
+                  furnaceOnOff: localFurnAction
                 }, (err, result) => {
                   if (err) {
                       console.log("Got a DB error in savePipeTemp");
@@ -222,14 +256,16 @@ connection.connect((err) => {
                   return;
                 }
               );
-              // end get flags from com controller
-    //        };
-            currentSaveDelayCount = saveDelayIntervalSeconds;
+              // end write temperatures to the database
+          currentSaveDelayCount = saveDelayIntervalSeconds;
+          localFurnAction = "noChange";
+          };
+
 //          };
 //          delayCount++;
         };
     // end if save delay < 21
-    }
+    //}
 //          currentSaveDelayCount = currentSaveDelayCount - 2;
   };
 
@@ -270,13 +306,6 @@ connection.connect((err) => {
     console.log("in dbcntrlr");
     console.log(newSettings);
 
-    // get existing settings and compare
-    //connection.query ("SELECT * FROM recirculatorSettings WHERE id=1", (err, result) => {
-    //  console.log("got back from getting recirc settings in update recirc setting");
-    //  console.log(result);
-    //  for (i=1; i < newSettings.length; i++) {
-    //    compare newSettings[i] to result[i]
-
     //  step through the newSettings object looking for a value to save
     for (var key in newSettings){
       if (newSettings[key] == ''){
@@ -299,7 +328,9 @@ connection.connect((err) => {
           if (err) {
             console.log("Got a DB error in update reg setting");
             console.log (err);
-          };
+          } else {
+            recircCNTRL.getNewRecircSettings();
+          }
         })
       }
     }
